Share glyphrun positioning and window creation
