+++ shell/Cargo.toml
@@ -15,11 +15,29 @@ cosmic-text = { workspace = true }
 cgmath = { workspace = true }
 
 [dev-dependencies]
+
+# markdown
+
 swash = { workspace = true }
 bytemuck = { workspace = true }
 env_logger = { workspace = true }
 derive_more = { workspace = true }
 inlyne = { workspace = true }
+
+# code (can't be compiled on wasm32 for now).
+
+[target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
+
+load-cargo = { path = "../rust-analyzer/crates/load-cargo" }
+project-model = { path = "../rust-analyzer/crates/project-model" }
+vfs = { path = "../rust-analyzer/crates/vfs" }
+base-db = { path = "../rust-analyzer/crates/base-db" }
+syntax = { path = "../rust-analyzer/crates/syntax" }
+# Semantics
+hir = { path = "../rust-analyzer/crates/hir" }
+# Syntax highlighting, AnalysisHost
+ide = { path = "../rust-analyzer/crates/ide" }
+
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 getrandom = { version = "0.2.12", features = ["js"] }
