--- shell/examples/code.rs
@@ -1,4 +1,0 @@
-#[path = "shared/application.rs"]
-mod application;
-
-fn main() {}
